added RoutPlanner implementation and changed stuff to doubles

+++ core/src/main/java/uk/ac/cam/teamdelta/Direction.java
@@ -2,10 +2,10 @@ package uk.ac.cam.teamdelta;
 
 public class Direction
 {
-	public Direction(float degrees)
+	public Direction(double degrees)
 	{
 		m_degrees = degrees;
 	}
-	public float getDegrees() {return m_degrees;}
-	private float m_degrees;
+	public double getDegrees() {return m_degrees;}
+	private double m_degrees;
 }
